match.c: Optimized find_match algorithm
This patch replaces the current implementation of find_match with
a faster implementation.

The faster implementation works by precomputing the match
conditions as a series of masks that match logic-high, low, change
and no-change. Because the match-terms are now combined into words,
there is now bit/word level parallelism; all the terms of a single
condition are combined in a single set of masks, which can be
matched to samples in all at once.

Further speed improvements are gained by...
 * ...eliminating linked-list traversal from the inner loop. The
   masks are stored contigiously within an array to maintain good
   cache coherence.
 * ...eliminating reduntant counters. Previously a loop-counter,
   abs_cur_samplenum and the term skip counters were incremented
   within the loop. These are now represented by a single
   incrementing pointer sample_pos. Other counters are updated
   once when the loop has finished.
 * ...removing sample position calculation in every loop. Instead
   the sample position pointerr is calculated once, and incremented
   by the unit-size every loop.
 * ...taking advantage of unaligned access, where available, to
   read whole sample words at once rather than loading words piece-
   wise in single bytes.