i2s: Replaced wav binary output with raw
The WAV binary format is a nice feature in principle, but in
practice the current implementation is broken and can't easily be
fixed.

The issue is that it is not possible to seek around in the output
binary stream, so there is no way to write the correct values for
certain fields in the WAV headers, particularly the data length,
once the audio has been fully decoded.

This limitation is not obvious to the user. Therefore, this patch
removes the WAV binary decode feature, and replaces it with a
simple RAW binary decode feature, which is less likely to confuse
the user.